services/data_refresher/requirements.txt
requests
web.py
google-cloud-storage
google-cloud-bigquery
pytrends